+++ src/FarNet.Redis/DB.cs
@@ -9,39 +9,18 @@ public static class DB
     const int DefaultIndex = -1;
 
     private static readonly ConcurrentDictionary<(string, int), IDatabase> s_databases = [];
-    private static IDatabase? s_database;
 
-    public static IDatabase? DefaultDatabase => s_database;
-
-    public static IDatabase OpenDefaultDatabase(Action<ConfigurationOptions>? configure = null)
-    {
-        if (s_database is { })
-            return s_database;
-
-        var configuration = Environment.GetEnvironmentVariable("FARNET_REDIS_CONFIGURATION")
-            ?? throw new InvalidOperationException("Requires environment variable FARNET_REDIS_CONFIGURATION.");
-
-        var options = ConfigurationOptions.Parse(configuration);
-        configure?.Invoke(options);
-
-        return s_database = Open(options);
-    }
+    public static string DefaultConfiguration =>
+        Environment.GetEnvironmentVariable("FARNET_REDIS_CONFIGURATION")
+        ?? throw new InvalidOperationException("Requires environment variable FARNET_REDIS_CONFIGURATION.");
 
     public static IDatabase Open(string configuration, int index = DefaultIndex)
     {
-        var options = ConfigurationOptions.Parse(configuration);
-        return Open(options, index);
-    }
-
-    public static IDatabase Open(ConfigurationOptions options, int index = DefaultIndex)
-    {
-        var key = options.ToString();
-        var db = s_databases.GetOrAdd((key, index), key =>
+        var db = s_databases.GetOrAdd((configuration, index), key =>
         {
-            var redis = ConnectionMultiplexer.Connect(options);
+            var redis = ConnectionMultiplexer.Connect(key.Item1);
             return redis.GetDatabase(key.Item2);
         });
-
         return db;
     }
 
@@ -56,9 +35,6 @@ public static class DB
 
     public static void Close(IDatabase db)
     {
-        if (ReferenceEquals(s_database, db))
-            s_database = null;
-
         db.Multiplexer.Close();
         foreach (var kv in s_databases.ToList())
         {